sweet alert

+++ src/pages/Management/Management.jsx
@@ -3,6 +3,8 @@ import { useDispatch } from 'react-redux';
 import './Management.scss';
 import { createGame,editGame,deleteGame } from "../../redux/games/games.actions";
 import { useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2'
+
 
 const INITIAL_STATE = {
   title: '',
@@ -21,11 +23,14 @@ function Management(props) {
 
 
   const [form, setForm] = useState(INITIAL_STATE);
+  const [formClean, setFormClean] = useState(INITIAL_STATE);
   const navigate = useNavigate();
 
   const handleInput = (event) => {
     const { name, value } = event.target;
     setForm({ ...form, [name]: value });
+    console.log('1',form);
+    console.log('2',INITIAL_STATE);
   };
 
   const changeCheckbox = (event) => {
@@ -47,24 +52,62 @@ function Management(props) {
   const dispatch = useDispatch();
   
   const submitGame = (event) =>{
+    let message ="";
 
     event.preventDefault();
     if(!form.title || !form.description || !form.trailer || !form.platform || !form.img) {
-      console.log('Some of the fields are not filled.');
+      Swal.fire('Some of the fields are not filled.');
       return;
     }
 
     if (form.id === ''){
       dispatch(createGame(form))
+      message="Game created";
     }else{
       dispatch(editGame(form))
+      message="Game edited";
     }
-    navigate('/');
+
+    Swal.fire({
+      position: 'top-end',
+      icon: 'success',
+      title: message,
+      showConfirmButton: false,
+      timer: 1500
+    })
+
+    setTimeout(() => {
+      navigate('/');
+    }, 600);
   }
 
   const deleteGameButton = () =>{
-    dispatch(deleteGame(form.id))
-    navigate('/');
+    
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+
+      dispatch(deleteGame(form.id))
+
+      Swal.fire({
+        position: 'top-end',
+        icon: 'success',
+        title: 'Game deleted',
+        showConfirmButton: false,
+        timer: 1500
+      })
+
+      setTimeout(() => {
+        navigate('/');
+      }, 600);
+    })
+
   }
 
   return (
@@ -176,20 +219,20 @@ function Management(props) {
           {form.id!=='' && <button type="button" onClick={deleteGameButton}> Delete </button>}
         </div>
       </form>
-      <ng-container>
-        <div className="card">
-          <img
+        {form !== formClean && <div className="card">
+          {form.img!=='' && <img
             src={ form.img }
             onError= {defaultPic}
             alt={ form.title }
-          />
+          />}
           <div className="card__content">
-            {form.title && <h3 className="title">{form.title}</h3>}
-            {form.platform && <h4>{form.platform.join(' - ')}</h4>} 
-            {form.genre && <h4 className="genre"> {form.genre}</h4>}
+            {form.title && <h3 className="title">Title: {form.title}</h3>}
+            {form.description && <h3 className="description">Description: {form.description}</h3>}
+            {form.platform && <h4>Platform: {form.platform.join(' - ')}</h4>} 
+            {form.genre && <h4 className="genre">Genre: {form.genre}</h4>}
+            {form.trailer && <h4 className="trailer">Trailer: {form.trailer}</h4>}
           </div>
-        </div>
-      </ng-container>
+        </div>}
     </div>
   </section>
 </>
